更新

+++ raspberry-pi/init.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
+set -e
 
-sudo cp /etc/apt/sources.list /etc/apt/sources.list.bak
+# 换apt源
+if [ -f "/etc/apt/sources.list.bak" ]; 
+then 
+    echo '/etc/apt/sources.list.bak备份存在'
+else 
+    sudo cp /etc/apt/sources.list /etc/apt/sources.list.bak
+fi
+
 sudo bash -c 'cat > /etc/apt/sources.list <<EOF
 deb https://mirrors.aliyun.com/debian/ bullseye main non-free contrib
 deb-src https://mirrors.aliyun.com/debian/ bullseye main non-free contrib
@@ -12,7 +20,6 @@ deb https://mirrors.aliyun.com/debian/ bullseye-backports main non-free contrib
 deb-src https://mirrors.aliyun.com/debian/ bullseye-backports main non-free contrib
 EOF
 '
-
-
 sudo apt-get update 
 sudo apt-get upgrade -y
+
